修改t2数据分析bug

+++ tencent-model/src/main/java/cc/pp/service/tencent/model/PicProporty.java
@@ -3,60 +3,60 @@ package cc.pp.service.tencent.model;
 import java.util.List;
 
 public class PicProporty {
-	
-	private List<Integer> pic_XDPI;
-	private List<Integer> pic_YDPI;
-	private List<Integer> pic_height;
-	private List<Integer> pic_size;
-	private List<Integer> pic_type;
-	private List<Integer> pic_width;
+
+	private List<Long> pic_XDPI;
+	private List<Long> pic_YDPI;
+	private List<Long> pic_height;
+	private List<Long> pic_size;
+	private List<Long> pic_type;
+	private List<Long> pic_width;
 	private List<String> url;
 
-	public List<Integer> getPic_XDPI() {
+	public List<Long> getPic_XDPI() {
 		return pic_XDPI;
 	}
 
-	public void setPic_XDPI(List<Integer> pic_XDPI) {
+	public void setPic_XDPI(List<Long> pic_XDPI) {
 		this.pic_XDPI = pic_XDPI;
 	}
 
-	public List<Integer> getPic_YDPI() {
+	public List<Long> getPic_YDPI() {
 		return pic_YDPI;
 	}
 
-	public void setPic_YDPI(List<Integer> pic_YDPI) {
+	public void setPic_YDPI(List<Long> pic_YDPI) {
 		this.pic_YDPI = pic_YDPI;
 	}
 
-	public List<Integer> getPic_height() {
+	public List<Long> getPic_height() {
 		return pic_height;
 	}
 
-	public void setPic_height(List<Integer> pic_height) {
+	public void setPic_height(List<Long> pic_height) {
 		this.pic_height = pic_height;
 	}
 
-	public List<Integer> getPic_size() {
+	public List<Long> getPic_size() {
 		return pic_size;
 	}
 
-	public void setPic_size(List<Integer> pic_size) {
+	public void setPic_size(List<Long> pic_size) {
 		this.pic_size = pic_size;
 	}
 
-	public List<Integer> getPic_type() {
+	public List<Long> getPic_type() {
 		return pic_type;
 	}
 
-	public void setPic_type(List<Integer> pic_type) {
+	public void setPic_type(List<Long> pic_type) {
 		this.pic_type = pic_type;
 	}
 
-	public List<Integer> getPic_width() {
+	public List<Long> getPic_width() {
 		return pic_width;
 	}
 
-	public void setPic_width(List<Integer> pic_width) {
+	public void setPic_width(List<Long> pic_width) {
 		this.pic_width = pic_width;
 	}
